requirements.txt
anyio==4.0.0
arrow==1.2.3
asgiref==3.7.2
binaryornot==0.4.4
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cookiecutter==2.3.0
Django==4.2.4
django-filter==23.2
django-model-utils==4.3.1
django-tailwind==3.6.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
jsonfield==3.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
Pygments==2.16.1
python-dateutil==2.8.2
python-slugify==8.0.1
python-telegram-bot==20.5
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rich==13.5.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
swapper==1.3.0
text-unidecode==1.3
tzdata==2023.3
urllib3==2.0.4
